requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-telegram-bot
nest_asyncio